requirements.txt
numpy>=1.20
netCDF4>=1.5
matplotlib>=3.5
cartopy>=0.20
scipy>=1.7
pyyaml>=5.4
